FC도착: 데모 명령·걷어내기 SQL·스타일가이드·smoke 한 바퀴, RLS·단계 갱신 버그 수정
- demo:seed/status/purge, db:migrate, admin:create (purge 는 SQL 파일만 생성)
- INSERT … RETURNING 이 자기 표 id 재조회 정책에 막히던 문제: 행 값 판정 함수로 교체 + 시험
- 단계 갱신 SQL 매개변수 형 충돌, 갱신 뒤 다음 단계 선택이 멈추던 문제
- Playwright smoke(DEMO on/off), verify 스크립트, pre-push hook(기본 꺼짐)
- 문서: DEPLOY·DEMO·DECISIONS·CLAUDE.md, .env.example(이름만)

+++ fcdochak/src/app/actions/partner.ts
@@ -180,8 +180,8 @@ export async function updateStage(input: z.infer<typeof StageInput>): Promise<Re
     if (d.returned != null && d.returned > s.units) return { error: '회송 수량이 선적 수량보다 많습니다' };
     await q.query(`insert into fcd.shipment_events (shipment_id, stage, raw_status, note, occurred_at, created_by) values ($1,$2,$3,$4,$5,$6)`, [d.shipmentId, d.stage, d.raw || null, d.note || null, at.toISOString(), v.id]);
     await q.query(
-      `update fcd.shipments set stage = $2, delivered_at = case when $2 = 9 then coalesce(delivered_at, $3::timestamptz) else delivered_at end,
-              fc_returned_units = coalesce($4, fc_returned_units) where id = $1`,
+      `update fcd.shipments set stage = $2::int, delivered_at = case when $2::int = 9 then coalesce(delivered_at, $3::timestamptz) else delivered_at end,
+              fc_returned_units = coalesce($4::int, fc_returned_units) where id = $1`,
       [d.shipmentId, d.stage, at.toISOString(), d.stage === 9 ? (d.returned ?? 0) : null],
     );
     if (d.stage === 9 && d.returned) {
